Improve error handling and warnings
These components back off from batch to single processing in order to
handle problematic docs that lead to errors for the transformers
`Pipeline` processing (known issues are with "empty" docs like `""` or
`" "` but there could potentially be other issues depending on the
model).

The warnings are a bit noisy, but I'd rather not crash on or skip entire
batches containing problematic texts, and if there is no indication,
then it's not clear to users why there can be huge speed differences
between individual batches.